integrate ahmm
