--- tatu_re/model/hmm/__init__.py
@@ -1,1 +1,0 @@
-from tatu_re.model.hmm.hmm import HMMStockPredictor
