+++ tatu_re/model/train.py
@@ -1,16 +1,22 @@
 from tatu_re.model.data_processing import calculate_target
 from tatu_re.model.linear_regression import linear_regression_pipeline
+from tatu_re.model.hmm import HMMStockPredictor
+from tatu_re.model.utils import save_model
 from tatu_re.utils import get_data
 
 from datetime import datetime
 import pathlib
 import pickle
 
-def train_model(save=True):
-
+def train_data():
     begin = datetime(2015,3,1)
     end = datetime(2021,3,1)
-    df = get_data(begin, end)
+    df = get_data(begin, end, "SPY")
+    return df
+
+def train_model(save=True):
+
+    df = train_data()
 
     target = calculate_target(df["Close"])
 
@@ -24,11 +30,15 @@ def train_model(save=True):
     model.fit(X, y)
 
     if save:
-        filename = 'linear_regression_v1.sav'
-        path = pathlib.Path(__file__).parent / "saved_models/"
-        if not path.exists(): 
-            path.mkdir()
-        path_to_save = path / filename
-        pickle.dump(model, open(path_to_save, 'wb'))
+        save_model(filename="linear_regression_v1.sav", model=model)
 
     return model
+
+def train_hidden_markov_model(save=True):
+    
+    # Initialise HMMStockPredictor object and fit the HMM
+    model = HMMStockPredictor(train_data=train_data())
+    model.fit()
+
+    if save:
+        save_model(filename="hmm_v1.sav", model=model)
